change u256 to u128

+++ src/VaultManager.sol
@@ -12,20 +12,20 @@ contract VaultManager is IVaultManager, LedgerComponent {
     // crossChainManagerAddress contract address
     address public crossChainManagerAddress;
     // valut balance, used for check if withdraw is valid
-    mapping(uint256 => mapping(bytes32 => uint256)) chain2symbol2balance;
+    mapping(uint256 => mapping(bytes32 => uint128)) chain2symbol2balance;
 
     // get balance
-    function getBalance(uint256 _chainId, bytes32 _symbol) public view override returns (uint256) {
+    function getBalance(uint256 _chainId, bytes32 _symbol) public view override returns (uint128) {
         return chain2symbol2balance[_chainId][_symbol];
     }
 
     // add balance
-    function addBalance(uint256 _chainId, bytes32 _symbol, uint256 _deltaBalance) public override onlyLedger {
+    function addBalance(uint256 _chainId, bytes32 _symbol, uint128 _deltaBalance) public override onlyLedger {
         chain2symbol2balance[_chainId][_symbol] += _deltaBalance;
     }
 
     // sub balance
-    function subBalance(uint256 _chainId, bytes32 _symbol, uint256 _deltaBalance) public override onlyLedger {
+    function subBalance(uint256 _chainId, bytes32 _symbol, uint128 _deltaBalance) public override onlyLedger {
         chain2symbol2balance[_chainId][_symbol] -= _deltaBalance;
     }
 }
